improve error message
