+++ src/api/app.ts
@@ -1,0 +1,23 @@
+import { DEFAULT_APP_NAME, tweaks } from "@config";
+import { axiod } from "@deps";
+
+export const getAppNameArt = async () => {
+  const url = "https://asciified.thelicato.io/api";
+
+  if (tweaks.appName) {
+    try {
+      const { data } = await axiod.get<string>(`${url}`, {
+        params: {
+          font: "ANSI Shadow",
+          text: tweaks.appName || "accli",
+        },
+        responseType: "text",
+      });
+
+      return data;
+    } catch {
+      /* Do nothing */
+    }
+  }
+  return DEFAULT_APP_NAME;
+};
