+++ src/lib/spinner.ts
@@ -1,0 +1,3 @@
+import { wait } from "@deps";
+
+export const spinner = wait("Downloading...⌛");
